src/requirements.txt
# Basic and production requirements
Django>=2.1.11,<2.2
djangorestframework==3.9.*
django-compressor==2.2.*
django-crispy-forms==1.7.*
django-filter==2.0.*
django-libsass
django-solo==1.1.*
markdown==2.6.*
qrcode==5.3
reportlab==3.4.*
requests==2.21.*

# Testing and development requirements
pytest
pytest-django
Faker==0.8.*
pytest-cov==2.*
PyYAML==5.1
isort
flake8
